+++ src/routes/queues/$id/index.tsx
@@ -1,26 +1,26 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { Suspense } from "react";
 import { z } from "zod";
-import { QueueSkeleton } from "@/components/skeletons/QueueSkeleton";
 import { checkAuthenticated } from "@/routes/__root";
-import { QueueControl } from "@/routes/queues/$id/-components/QueueControl";
+import { Queue } from "@/routes/queues/$id/-components/Queue";
+import { QueueSkeleton } from "@/routes/queues/$id/-components/QueueSkeleton";
 
 const queueSearchSchema = z.object({
     code: z.string().optional(),
 });
 
 export const Route = createFileRoute("/queues/$id/")({
-    component: Queue,
+    component: QueueLoader,
     validateSearch: (search: Record<string, unknown>) => {
         return queueSearchSchema.parse(search);
     },
     beforeLoad: async ({ location }) => await checkAuthenticated(location.href),
 });
 
-function Queue() {
+function QueueLoader() {
     return (
         <Suspense fallback={<QueueSkeleton />}>
-            <QueueControl />
+            <Queue />
         </Suspense>
     );
 }
